Add skeleton for each page on the site
Yeah so I didn't really like the class names that I gave
to all of the content, it was making my head hurt. So I
rebuilt the skeleton and tried to use naming conventions
that made sense. Each page is going to look the same,
pretty much, with the only real difference being the main
content. I've gotta style the header and footer to my liking
but this should provide a more consistent base to work on,
especially since I'll be working with someone else on this.